# pims/pims/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: Cytomine Python Image Managament Server PIMS
  description: |
    Cytomine Python Image Managament Server (PIMS) HTTP API. While this API is intended to be internal, a lot of the
    following specification can be ported to the external Cytomine API.

    # Additional API formats
    In addition to OpenAPI types and formats, the following table contains some additional formats used in this
    specification.

    | Type | Format | Examples | Specification |
    |------|--------|----------|---------------|
    | `string` | `range` | **"1:3"**, **"2:"**, **":"**| Inspired from Python slicing |
    | `string` | `wkt` | **"POINT (10 20)"** | [Open Geospacial Consortium](https://www.ogc.org/standards/sfa)|
    | `string` | `color` | **"#FF00AA"**, **"rgb(100,0,0)"**, **"red"** | [CSS Color Module Level 3](https://drafts.csswg.org/css-color-3/#colorunits)|
    | `string` | `colormap-id` | **"jet"**, **"!jet"**||
    | `string` | `filter-id` | **"yen"** ||
  version: 0.1.0
#servers:
#  - url: 'https'
...
paths:
  /image/{filepath}/normalized-tile/zoom/{zoom}/ti/{ti}:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/path-zoom'
      - $ref: '#/components/parameters/path-tile-index'
      - $ref: '#/components/parameters/header-etag'
    get:
      operationId: pims.api.tile.show_normalized_tile_by_zoom
      summary: Get a normalized tile at zoom
      description: |
        Get a 8-bit normalized tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        tile is extracted from the median focal plane at first timepoint.
      tags:
        - Normalized tiles
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-min-intensities-visualisation'
        - $ref: '#/components/parameters/query-max-intensities-visualisation'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
    post:
      operationId: pims.api.tile.show_normalized_tile_with_body_by_zoom
      summary: Get a normalized tile at zoom
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a 8-bit normalized tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Normalized tiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/window-plane-selection'
                - $ref: '#/components/schemas/image-manipulation-visualisation'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
  /image/{filepath}/normalized-tile/level/{level}/ti/{ti}:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/path-level'
      - $ref: '#/components/parameters/path-tile-index'
      - $ref: '#/components/parameters/header-etag'
    get:
      operationId: pims.api.tile.show_normalized_tile_by_level
      summary: Get a normalized tile at level
      description: |
        Get a 8-bit normalized tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Normalized tiles
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-min-intensities-visualisation'
        - $ref: '#/components/parameters/query-max-intensities-visualisation'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
    post:
      operationId: pims.api.tile.show_normalized_tile_with_body_by_level
      summary: Get a normalized tile at level
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a 8-bit normalized tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Normalized tiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/window-plane-selection'
                - $ref: '#/components/schemas/image-manipulation-visualisation'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
...
  /image/{filepath}/tile/zoom/{zoom}/ti/{ti}:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/path-zoom'
      - $ref: '#/components/parameters/path-tile-index'
      - $ref: '#/components/parameters/header-etag'
    get:
      operationId: pims.api.tile.show_tile_by_zoom
      summary: Get a tile at zoom
      description: |
        Get a 8-bit tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        tile is extracted from the median focal plane at first timepoint.
      tags:
        - Tiles
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-min-intensities-visualisation'
        - $ref: '#/components/parameters/query-max-intensities-visualisation'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
    post:
      operationId: pims.api.tile.show_tile_with_body_by_zoom
      summary: Get a tile at zoom
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a 8-bit tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Tiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/window-plane-selection'
                - $ref: '#/components/schemas/image-manipulation-visualisation'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
  /image/{filepath}/tile/level/{level}/ti/{ti}:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/path-level'
      - $ref: '#/components/parameters/path-tile-index'
      - $ref: '#/components/parameters/header-etag'
    get:
      operationId: pims.api.tile.show_tile_by_level
      summary: Get a tile at level
      description: |
        Get a 8-bit tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Tiles
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-min-intensities-visualisation'
        - $ref: '#/components/parameters/query-max-intensities-visualisation'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
    post:
      operationId: pims.api.tile.show_tile_with_body_by_level
      summary: Get a tile at level
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a 8-bit tile at a given zoom level and tile index, optimized for visualisation, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Tiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/window-plane-selection'
                - $ref: '#/components/schemas/image-manipulation-visualisation'
      responses:
        200:
          $ref: '#/components/responses/tile'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
...
components:
  parameters:
    header-etag:
      name: ETag
      in: header
      description: |
        The RFC 7232 ETag header field in a response provides the entity-tag of
        a selected resource. The entity-tag is an opaque identifier for versions
        and representations of the same resource over time, regardless whether
        multiple versions are valid at the same time. An entity-tag consists of
        an opaque quoted string, possibly prefixed by a weakness indicator (see
        [RFC 7232 Section 2.3](https://tools.ietf.org/html/rfc7232#section-2.3)).
      required: false
      schema:
        type: string
...
    path-filepath:
      name: filepath
      in: path
      description: The file path, relative to server base path.
      required: true
      schema:
        type: string
        format: path
      example: '123/my-file.ext'
    path-zoom:
      name: zoom
      in: path
      description: |
        The zoom level.
        Zoom 0 denotes the tier with the worst resolution (smallest image, top of the image pyramid).
        Maximum admissible zoom level depends on image.
      required: true
      schema:
        $ref: '#/components/schemas/zoom-or-level'
...
    path-tile-index:
      name: ti
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/tile-index'
...
    query-channel-selection:
      name: channels
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-channel-index'
...
    query-z-slice-selection:
      name: z_slices
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-z-slice-index'
...
    query-timepoint-selection:
      name: timepoints
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-timepoint-index'
...
    query-c-reduction:
      name: c_reduction
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/channel-reduction'
    query-z-reduction:
      name: z_reduction
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/z-slice-reduction'
...